Create docs.

+++ src/services/dao/mongo/product.dao.js
@@ -7,6 +7,7 @@ export default class ProductDAO {
     try {
       return await productModel.create(product);
     } catch (error) {
+      console.log(error);
       throw new Error("Product not created");
     }
   };
@@ -106,7 +107,7 @@ export default class ProductDAO {
     return await productModel.updateOne({ _id: id }, product);
   };
 
-  deleteProduct = async (id) => {
-    return await productModel.deleteOne({ _id: id });
+  deleteProduct = async (id, userId) => {
+    return await productModel.deleteOne({ _id: id, seller: userId });
   };
 }
